# Re-implement all offline dictionary based on the fixed byte value reader