--- a/pinot-core/src/main/java/com/linkedin/pinot/core/io/writer/impl/v1/FixedBitSingleValueWriter.java
+++ b/pinot-core/src/main/java/com/linkedin/pinot/core/io/writer/impl/v1/FixedBitSingleValueWriter.java
@@ -16,21 +16,21 @@
 package com.linkedin.pinot.core.io.writer.impl.v1;
 
 import com.linkedin.pinot.common.segment.ReadMode;
+import com.linkedin.pinot.core.io.util.FixedBitIntReaderWriter;
 import com.linkedin.pinot.core.io.writer.SingleColumnSingleValueWriter;
-import com.linkedin.pinot.core.io.writer.impl.FixedBitIntWriter;
 import com.linkedin.pinot.core.segment.memory.PinotDataBuffer;
 import java.io.File;
 import java.nio.channels.FileChannel;
 
 
 public class FixedBitSingleValueWriter implements SingleColumnSingleValueWriter {
-  private FixedBitIntWriter dataFileWriter;
+  private FixedBitIntReaderWriter dataFileWriter;
 
   public FixedBitSingleValueWriter(File file, int rows, int columnSizeInBits) throws Exception {
     PinotDataBuffer dataBuffer =
         PinotDataBuffer.fromFile(file, 0, (rows * columnSizeInBits + Byte.SIZE - 1) / Byte.SIZE, ReadMode.mmap,
             FileChannel.MapMode.READ_WRITE, file.getAbsolutePath());
-    dataFileWriter = new FixedBitIntWriter(dataBuffer, rows, columnSizeInBits);
+    dataFileWriter = new FixedBitIntReaderWriter(dataBuffer, rows, columnSizeInBits);
   }
 
   @Override