--- a/pinot-core/src/main/java/com/linkedin/pinot/core/realtime/impl/dictionary/MutableDictionary.java
+++ b/pinot-core/src/main/java/com/linkedin/pinot/core/realtime/impl/dictionary/MutableDictionary.java
@@ -16,55 +16,15 @@
 
 package com.linkedin.pinot.core.realtime.impl.dictionary;
 
-import com.linkedin.pinot.core.segment.index.readers.Dictionary;
+import com.linkedin.pinot.core.segment.index.readers.BaseDictionary;
 import javax.annotation.Nonnull;
 
 
-public abstract class MutableDictionary implements Dictionary {
+public abstract class MutableDictionary extends BaseDictionary {
 
- @Override
+  @Override
   public String getStringValue(int dictId) {
     return get(dictId).toString();
-  }
-
-  @Override
-  public void readIntValues(int[] dictIds, int startPos, int limit, int[] outValues, int outStartPos) {
-    int endPos = startPos + limit;
-    for (int i = startPos; i < endPos; i++) {
-      outValues[outStartPos++] = getIntValue(dictIds[i]);
-    }
-  }
-
-  @Override
-  public void readLongValues(int[] dictIds, int startPos, int limit, long[] outValues, int outStartPos) {
-    int endPos = startPos + limit;
-    for (int i = startPos; i < endPos; i++) {
-      outValues[outStartPos++] = getLongValue(dictIds[i]);
-    }
-  }
-
-  @Override
-  public void readFloatValues(int[] dictIds, int startPos, int limit, float[] outValues, int outStartPos) {
-    int endPos = startPos + limit;
-    for (int i = startPos; i < endPos; i++) {
-      outValues[outStartPos++] = getFloatValue(dictIds[i]);
-    }
-  }
-
-  @Override
-  public void readDoubleValues(int[] dictIds, int startPos, int limit, double[] outValues, int outStartPos) {
-    int endPos = startPos + limit;
-    for (int i = startPos; i < endPos; i++) {
-      outValues[outStartPos++] = getDoubleValue(dictIds[i]);
-    }
-  }
-
-  @Override
-  public void readStringValues(int[] dictIds, int startPos, int limit, String[] outValues, int outStartPos) {
-    int endPos = startPos + limit;
-    for (int i = startPos; i < endPos; i++) {
-      outValues[outStartPos++] = getStringValue(dictIds[i]);
-    }
   }
 
   public abstract void index(@Nonnull Object rawValue);