Also fixed the OnHeapStringDictionary.indexOf() to return correct value when value does not exist (Should be the same as Arrays.binarySearch() to make range filter work properly)

Most performance gain comes from StringDictionary for the following reasons:
  1. No need to duplicate ByteBuffer, index byte-by-byte instead
  2. Reuse the byte buffer if possible
  3. When reading unpadded string, no need to read all bytes

Benchmark Results:
Old Dictionaries:
Benchmark                                     Mode  Cnt      Score      Error  Units
BenchmarkOfflineIndexReader.doubleDictionary  avgt    3   1131.025 ±   54.210  us/op
BenchmarkOfflineIndexReader.floatDictionary   avgt    3   1080.044 ±   16.400  us/op
BenchmarkOfflineIndexReader.intDictionary     avgt    3   1085.413 ±   61.710  us/op
BenchmarkOfflineIndexReader.longDictionary    avgt    3    901.671 ±    9.518  us/op
BenchmarkOfflineIndexReader.stringDictionary  avgt    3  16020.946 ± 1452.813  us/op

New Dictionaries:
Benchmark                                     Mode  Cnt     Score     Error  Units
BenchmarkOfflineIndexReader.doubleDictionary  avgt    3   985.736 ±   5.780  us/op
BenchmarkOfflineIndexReader.floatDictionary   avgt    3   939.546 ±  10.791  us/op
BenchmarkOfflineIndexReader.intDictionary     avgt    3   958.593 ±  33.669  us/op
BenchmarkOfflineIndexReader.longDictionary    avgt    3   769.179 ±  38.159  us/op
BenchmarkOfflineIndexReader.stringDictionary  avgt    3  8990.013 ± 256.303  us/op